tools/mcp-task-server/pyproject.toml
[project]
name = "fuellhorn-task-mcp"
version = "0.1.0"
description = "MCP Server für GitHub Issue Task Management"
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.0.0",
    "PyGithub>=2.1.0",
]

[project.scripts]
fuellhorn-tasks = "server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["server.py"]
